+++ src/components/CartDescription/style.tsx
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.div`
+  margin-bottom: 36px;
+`;
