+++ src/components/common/Button/Button.tsx
@@ -1,0 +1,8 @@
+import { Wrapper } from "./style";
+import { ButtonProps } from "../../../types/button";
+
+const Button = ({ children, ...rest }: ButtonProps) => {
+  return <Wrapper {...rest}>{children}</Wrapper>;
+};
+
+export default Button;
