+++ src/components/index.ts
@@ -1,9 +1,9 @@
-import CartDescription from "./CartDescription";
-import CartItemList from "./CartItemList";
-import Header from "./Header";
-import Footer from "./Footer";
-import CartPrice from "./CartPrice";
-import OrderConfirm from "./OrderConfirm";
+import CartDescription from "./CartDescription/CartDescription";
+import CartItemList from "./CartItemList/CartItemList";
+import Header from "./Header/Header";
+import Footer from "./Footer/Footer";
+import CartPrice from "./CartPrice/CartPrcie";
+import OrderConfirm from "./OrderConfirm/OrderConfirm";
 
 export {
   CartDescription,
