+++ src/layouts/index.ts
@@ -1,0 +1,3 @@
+import CartLayout from "./CartLayout/CartLayout";
+
+export { CartLayout };
